Implemented smooth transitions for wind speed and direction changes.
Users can now set the wind speed and direction (changed to an angle instead of a vector) and expect the clouds to smoothly transition to the new state. On the CPU side we now keep state and integrate all cloud-related positions over time, sending the results of those calculations to the shader as part of the constants. Those were slightly refactored, dropping parameters that were no longer needed and adding a few new ones. That in turn affected the padding. The compute shader was adjusted accordingly, removing some bits of math that were no longer needed.