Moved all frame based calculations.
All calculations that change parameters in frame_data are now moved to the _update_per_frame_data function.
In _fill_push_constant I am now using the _time that is stored in the frame_data instead of one that was obtained in this function to fix a bug in the existing code that would cause "seams" to show in the sky dome at high wind speeds because time was updated too often.