Implemented smooth transitions for wind speed and direction changes. (#6)
* Implemented smooth transitions for wind speed and direction changes.

Users can now set the wind speed and direction (changed to an angle instead of a vector) and expect the clouds to smoothly transition to the new state. On the CPU side we now keep state and integrate all cloud-related positions over time, sending the results of those calculations to the shader as part of the constants. Those were slightly refactored, dropping parameters that were no longer needed and adding a few new ones. That in turn affected the padding. The compute shader was adjusted accordingly, removing some bits of math that were no longer needed.

* Moved all frame based calculations.

All calculations that change parameters in frame_data are now moved to the _update_per_frame_data function.
In _fill_push_constant I am now using the _time that is stored in the frame_data instead of one that was obtained in this function to fix a bug in the existing code that would cause "seams" to show in the sky dome at high wind speeds because time was updated too often.