Added date, time and location as well as a realistic sun position calculation based on that.
A new script is introduced that keeps track of the local date and time as well as the location on earth.
The sun script was enhanced with a calculation of the sun's position based on this new script.
The demo scene was enhanced to include this new script.
The cloud sky script was slightly tweaked to also hold a reference to the date, time and location which adds itself in a similar way as the sun script does.