Updates the performance parameters in the editor. (#8)
First change is to extract an update_performance() method out of most of what was in delayed_init() except for connecting the update_sky() function to the "frame_pre_draw".
Then setters were added to the two performance settings to make sure we do a full sky init and then an update of those performance parameters.
Note that I still see parts of the sky "flashing in red or green" the moment I make such a change.